refactor($numbers): move test to domain package
